extract metrics-jmx
