+++ metrics-core/src/test/java/io/dropwizard/metrics/ClockTest.java
@@ -10,18 +10,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.offset;
 
 public class ClockTest {
-    @Test
-    public void cpuTimeClock() throws Exception {
-        final Clock.CpuTimeClock clock = new Clock.CpuTimeClock();
-
-        assertThat((double) clock.getTime())
-                .isEqualTo(System.currentTimeMillis(),
-                           offset(100.0));
-
-        assertThat((double) clock.getTick())
-                   .isEqualTo(ManagementFactory.getThreadMXBean().getCurrentThreadCpuTime(),
-                              offset(1000000.0));
-    }
 
     @Test
     public void userTimeClock() throws Exception {
