+++ metrics-jmx/src/main/java/io/dropwizard/metrics/jmx/DefaultObjectNameFactory.java
@@ -1,7 +1,9 @@
-package io.dropwizard.metrics;
+package io.dropwizard.metrics.jmx;
 
 import javax.management.MalformedObjectNameException;
 import javax.management.ObjectName;
+
+import io.dropwizard.metrics.MetricName;
 
 public class DefaultObjectNameFactory implements ObjectNameFactory {
