+++ metrics-jmx/src/test/java/io/dropwizard/metrics/jmx/JmxReporterTest.java
@@ -1,4 +1,4 @@
-package io.dropwizard.metrics;
+package io.dropwizard.metrics.jmx;
 
 import org.junit.After;
 import org.junit.Before;
@@ -7,14 +7,13 @@ import org.junit.Test;
 import io.dropwizard.metrics.Counter;
 import io.dropwizard.metrics.Gauge;
 import io.dropwizard.metrics.Histogram;
-import io.dropwizard.metrics.JmxReporter;
 import io.dropwizard.metrics.Meter;
 import io.dropwizard.metrics.MetricFilter;
 import io.dropwizard.metrics.MetricName;
 import io.dropwizard.metrics.MetricRegistry;
-import io.dropwizard.metrics.ObjectNameFactory;
 import io.dropwizard.metrics.Snapshot;
 import io.dropwizard.metrics.Timer;
+import io.dropwizard.metrics.jmx.JmxReporter;
 
 import javax.management.*;
