+++ pom.xml
@@ -36,6 +36,7 @@
         <module>metrics-servlet</module>
         <module>metrics-servlets</module>
         <module>metrics-adapter</module>
+        <module>metrics-jmx</module>
     </modules>
 
     <properties>
